--lazy: don't commit if nothing has changed
getnear will remember what you told it to do.  If you run it again with the
--lazy option and nothing has changed, it won't do anything.

This is a band-aid for the fact that T-series updates take up to a minute.  A
better fix would be to make the code actually run quickly by not changing
things that don't need changing!