Fix: access ports can carry trunked traffic
The previous model required that a port either:

* be a "trunk" only port and carry all-trunked traffic (traffic that is
  already tagged with a VLAN id); or
* be an "access" only port, where inbound traffic is expected to be
  untagged, and a VLAN tag with that port's PVID is added once inside
  the switch.

This model was wrong.  It's completely fine to have a port that is both
an access port and a trunk port:

```python
ports = [
    {1: A},             #1: Upstream switch
    {1: A, 15: T},      #2: My laptop
    {1: A},
    {1: A},
    {15: A},            #5: Weird device that I'm isolating
    {1: A},
    {1: A},
    {1: A},
]
```

My laptop has a regular ethernet interface whose traffic is tagged into
VLAN 1 by the switch.  I've also asked Linux to add a VLAN 15 interface
as well, and the switch should accept that as trunked traffic.  (This is
so I can talk to some dodgy device on port 5.)